Generic service and generic stub refactor.
See https://github.com/grpc/proposal/pull/426
