+++ gRPC-C++.podspec
@@ -103,7 +103,9 @@ Pod::Spec.new do |s|
                       'include/grpcpp/ext/health_check_service_server_builder_option.h',
                       'include/grpcpp/ext/server_metric_recorder.h',
                       'include/grpcpp/generic/async_generic_service.h',
+                      'include/grpcpp/generic/callback_generic_service.h',
                       'include/grpcpp/generic/generic_stub.h',
+                      'include/grpcpp/generic/generic_stub_callback.h',
                       'include/grpcpp/grpcpp.h',
                       'include/grpcpp/health_check_service_interface.h',
                       'include/grpcpp/impl/call.h',
@@ -160,6 +162,7 @@ Pod::Spec.new do |s|
                       'include/grpcpp/impl/completion_queue_tag.h',
                       'include/grpcpp/impl/create_auth_context.h',
                       'include/grpcpp/impl/delegating_channel.h',
+                      'include/grpcpp/impl/generic_stub_internal.h',
                       'include/grpcpp/impl/grpc_library.h',
                       'include/grpcpp/impl/intercepted_channel.h',
                       'include/grpcpp/impl/interceptor_common.h',
